refactor(web): move exportSnapshot to SnapshotController
